# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Trenda API
  description: >
    Design & Development major growth platform.
    Includes project management, portfolio versioning,
    skill analysis, trend interaction, and AI-driven insights.
  version: 1.0.0
servers:
  - url: https://api.trenda.app
    description: Production
  - url: http://localhost:8080
    description: Local
...
paths:
...
  /api/users:
    get:
      tags: [User]
      summary: Get all users
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

    post:
      tags: [User]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/search:
    get:
      tags: [User]
      summary: Find user by email
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/{id}:
    get:
      tags: [User]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # Profile APIs
...
  /api/play/records/user/{userId}:
    get:
      tags: [Play]
      summary: Get user play records
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Play record list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        major:
          type: string
        createdAt:
          type: string
          format: date-time
...